LETS GET THIS PARTY STARTED IN HERE The page actually resembles how I think it'll look by the end?!?
After going through idea after Idea, I finally found an Idea that would be unique and cool. An interactive map of rail routes and shit.
So I made that, and put it up on my local environment to realize that it sucked. I don't know if I am just really bad at making train routes, or if that's just simply not the way to go, but I instead just did my classic thing:

Minimal UI, with an image in the background, sort of resembling the drawing I've made just now.